refactor: 페이지 데이터 fetching 커스텀 훅 분리
